+++ src/main/java/com/typelead/gradle/eta/internal/DefaultEtaConfiguration.java
@@ -3,7 +3,6 @@ package com.typelead.gradle.eta.internal;
 import java.io.File;
 import java.util.ArrayList;
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.HashSet;
 import java.util.LinkedHashSet;
 import java.util.List;